Combine sshd_config_dict with host_vars to merge those dictionaries
Unfortunately in Ansible (YAML, I guess) dictionaries replace one another
unless you combine them with a macro. In this cause it's useful to do this
so that we don't need to re-define the entire hash over again in hyperion's
hostvars.

The drawback is that you effectively need to separate the original default
plus the new settings out into separately named variables so that you can
use combine() on the final variable (the one used in the play)

 - In group_vars/all: change sshd_config_dict to point to
   sshd_config_dict_defaults, where config_config_dict_defaults
   holds the default values for the play (since non-roles don't have defaults)

 - In host_vars/hyperion: re-define only the necessary properites, and
   override sshd_config_dict by combining sshd_config_dict_defaults and
   sshd_config_dict_hyperion